feat: add OIDC_AUTH_URL and AFTER_LOGIN_URL env variables
These variables are required by the OIDC Authservice to correctly redirect users along the
authentication flow.
* OIDC_AUTH_URL: is used as the URL tht Authservice will hit for initiating the auth flow
* AFTER_LOGIN_URL: is used for redirecting after login
These variables are required because the default behaviour of OIDC Authservice is to redirect to
URLs that can be undesired, by setting them up, we have a deterministic way of setting where users
should be redirected.

Part of canonical/bundle-kubeflow#608